# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Bancaire Laravel",
        "description": "Documentation complète de l'API bancaire avec gestion des comptes, transactions et blocages automatiques",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "{L5_SWAGGER_CONST_HOST}",
            "description": "Serveur de production"
        },
        {
            "url": "http://localhost:8000",
            "description": "Serveur de développement"
        }
    ],
    "paths": {
        "/api/v1/comptes": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Lister les comptes bancaires",
                "description": "Récupère la liste paginée des comptes non archivés avec possibilité de filtrage par type, statut et tri",
                "operationId": "listerComptes",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page (défaut: 1)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Nombre d'éléments par page (défaut: 10, max: 100)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1,
                            "example": 10
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filtrer par type de compte",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "epargne",
                                "courant"
                            ],
                            "example": "epargne"
                        }
                    },
                    {
                        "name": "statut",
                        "in": "query",
                        "description": "Filtrer par statut du compte",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "actif",
                                "bloque"
                            ],
                            "example": "actif"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Champ de tri",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "dateCreation",
                                "numero",
                                "solde"
                            ],
                            "example": "dateCreation"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Ordre de tri",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ],
                            "example": "desc"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des comptes récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Compte"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/comptes/{numero}": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Détail d'un compte bancaire",
                "description": "Récupère les informations détaillées d'un compte par son numéro unique",
                "operationId": "detailCompte",
                "parameters": [
                    {
                        "name": "numero",
                        "in": "path",
                        "description": "Numéro du compte (format: CPT suivi de 13 chiffres)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "pattern": "^CPT[0-9]{13}$",
                            "example": "CPT0000000000001"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du compte récupérés avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Compte"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Compte non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Ressource non trouvée"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/comptes/client/{telephone}": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Comptes d'un client",
                "description": "Récupère tous les comptes d'un client via son numéro de téléphone",
                "operationId": "comptesParClient",
                "parameters": [
                    {
                        "name": "telephone",
                        "in": "path",
                        "description": "Numéro de téléphone du client",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "+221771234567"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Nombre d'éléments par page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comptes du client récupérés",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Compte"
                                            }
                                        },
                                        "pagination": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Client non trouvé ou aucun compte"
                    }
                }
            }
        },
        "/api/v1/comptes/{compteId}/bloquer": {
            "post": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Bloquer un compte bancaire",
                "description": "Permet de bloquer un compte épargne avec des dates de début et fin optionnelles. Seuls les comptes épargne peuvent être bloqués.",
                "operationId": "bloquerCompte",
                "parameters": [
                    {
                        "name": "compteId",
                        "in": "path",
                        "description": "ID du compte à bloquer",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Informations de blocage du compte",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "motif"
                                ],
                                "properties": {
                                    "motif": {
                                        "description": "Motif du blocage",
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Suspicion de fraude"
                                    },
                                    "date_debut_blockage": {
                                        "description": "Date de début du blocage (optionnel, défaut = maintenant)",
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-10-30 14:00:00"
                                    },
                                    "date_fin_blockage": {
                                        "description": "Date de fin du blocage (optionnel)",
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-11-30 14:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Compte bloqué avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Compte bloqué avec succès"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "version": {
                                            "type": "string",
                                            "example": "1.0.0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur de validation ou compte non éligible au blocage",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Seuls les comptes épargne peuvent être bloqués."
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "version": {
                                            "type": "string",
                                            "example": "1.0.0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Compte non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Ressource non trouvée"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "version": {
                                            "type": "string",
                                            "example": "1.0.0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Compte": {
                "title": "Compte bancaire",
                "description": "Objet représentant un compte bancaire",
                "properties": {
                    "id": {
                        "description": "ID unique du compte",
                        "type": "string",
                        "format": "uuid"
                    },
                    "numero": {
                        "description": "Numéro du compte (format: CPTXXXXXXXXXXXXX)",
                        "type": "string"
                    },
                    "solde": {
                        "description": "Solde du compte",
                        "type": "number",
                        "format": "decimal"
                    },
                    "type": {
                        "description": "Type de compte",
                        "type": "string",
                        "enum": [
                            "epargne",
                            "courant"
                        ]
                    },
                    "statut": {
                        "description": "Statut du compte",
                        "type": "string",
                        "enum": [
                            "actif",
                            "bloque",
                            "archive"
                        ]
                    },
                    "motifBlocage": {
                        "description": "Motif du blocage si applicable",
                        "type": "string",
                        "nullable": true
                    },
                    "date_debut_blockage": {
                        "description": "Date de début du blocage",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "date_fin_blockage": {
                        "description": "Date de fin du blocage",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "client_id": {
                        "description": "ID du client propriétaire",
                        "type": "string",
                        "format": "uuid"
                    },
                    "created_at": {
                        "description": "Date de création",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date de dernière modification",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Comptes",
            "description": "Gestion des comptes bancaires"
        }
    ]
}